requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
tensorflow>=2.6.0
cryptography>=3.4.7
phe>=1.4.0
mpy>=0.1.0
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
pytest>=6.2.5
black>=21.7b0
isort>=5.9.3
mypy>=0.910 